sdk_extra_vars.bbclass: restore bbclass
This class is used to add additional variable definitions to the
environment-setup file in an SDK directly, rather than through
environment-setup.d files. The latter is often the best solution, but at times
this is needed.

This was previously moved to meta-mentor-private, but we are now restoring it
for use in our SDK construction.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>