sokol-flex-support:96boards-tools: update to latest revision
With udev primitives changing in systemd the resize-helper script fails to
identify the root device correctly and hence the partition is not extended
successfully on boot.
The latest revision of the resize-helper from 96boards-tools takes care of this
change and so we pull that in to fix disk resizing on first boot. However, we
have to patch it to make it work with the version of parted we have in use that
requires a confirmation if the device is mounted.

JIRA: https://jira.alm.mentorg.com/browse/SB-19819

Signed-off-by: Awais Belal <awais_belal@mentor.com>