staging/udev-extraconf: include fix for systemd automount issue from upstream
This adds fix for systemd automount which has been fixed differently in
upstream oe-core layer.

Ref: https://github.com/openembedded/openembedded-core/commit/a3c93ec301a34413f91e3edb70c16454ebcdcdf2

With systemd-udev.service, the kernel and control sockets for systemd-udev
fails and hence systemd-udevd service fails to start which hangs the boot if
weston is enabled. This issue has been addressed upsrtream and fix is being included
here.

TODO: All the changes in flex-staging layer for udev-extraconf has been merged
upstream but will be part of 4.0.3 release which is planned to be out in Sep 2022.
For the time being, we are including the fix here but all these staging layer
changes should be dropped from here once 4.0.3 is out and we shift to that release.

JIRA-ID: SB-19773

Signed-off-by: ansar-rasool <ansar_rasool@mentor.com>