production-image: fix handling of incompatible license exceptions
We switched from our own image sanity check for install of incompatibly licensed
packages to the oe-core one, but as a result there's no variable to explicitly
disallow or allow this for production vs development images. So, rework it to
explicitly clear the exceptions by default in our production image.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>