sokol-flex: default VENDOR based on MACHINE layer name
Determine the default vendor based on the layer name of the layer which provides
the MACHINE .conf, with meta- removed. For example, for icicle-kit-es,
`conf/vendor/polarfire-soc-yocto-bsp.conf` would be included by default. Either
VENDOR or VENDOR_CONF may be overridden by a MACHINE, as will be the case for
our official Sokol Flex BSP machines.

JIRA: SB-19826

Signed-off-by: Christopher Larson <chris_larson@mentor.com>