image-root-password.bbclass: fix default empty root password login
The root user login on development-image and production-image does
not work when ROOT_PASSWORD is not set i.e. by default. This is due
to a bug in the image-root-password.bbclass function encrypt_root_pw
that does an invalid comparison of the password string with 'None'.
It should rather compare it with an empty string as that is the
default state of the ROOT_PASSWORD var.
Otherwise, the ACTUAL_ROOT_PASSWORD evaluates to "'*0'" which is not
a proper encrypted password.

JIRA: SB-20967

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>