flex-checkout: default the git init branch to main, not $branch
By prematurely setting our branch, the later logic to check out that
branch does nothing, as we're already on it. The problem is we never
actually did a checkout, so no files exist. Explicitly specify a default
branch which does not match our own to avoid skipping the later
checkout.

JIRA: SB-21205

Signed-off-by: Christopher Larson <chris_larson@mentor.com>