flex-checkout: fix for 'sokol-flex' and +snapshot
Due to the use of '-' as the separator in our manifest filenames, we
have to know precisely how to get the machine name back out of it, which
means the number of dashes in the DISTRO is a factor. Now that we're
`sokol-flex` rather than `mel`, we need to adjust the cut to the 4th,
not 3rd dash, and also strip off the +snapshot- suffix.

JIRA: SB-21211

Signed-off-by: Christopher Larson <chris_larson@mentor.com>