psplash: disable startup message
By default startup message is enabled. This draws a stripe close to the
bottom of the screen with the background color specified(currently black).

This might not be visible in some cases, where the image is centered and
the remaining area is filled with the background color.

Disable startup message to avoid the black rectangle displayed over the
psplash image.

Signed-off-by: Vijai Kumar K <vijaikumar.kanagarajan@siemens.com>