flex_utils.py/dump-licenses: rework file operations
Current implementation is opening the file at 3 differenct locations but
the same results can be achieved if we create an empty set and perform
all the operations on it and then write it to the file only once.
However, if append flag is set then a read operation is required prior
to writing to the file to remove duplicates.

JIRA: SB-22425

Signed-off-by: Tariq Mehmood <tariq.mehmood.ext@siemens.com>