flex_utils.py/dump-licenses: replace PN with BPN
Add source name to package lists for SVM monitoring list automation by
using BPN instead of PN. The PN variable can not serve this purpose
because in many cases it contains redundant suffixes and prefixes in
context of monitoring list, for example, -native, -cross, lib32-, and
nativesdk-. So, BPN gives us the base package name which is as close to
source name as possible in most cases and it also removes duplications
in the final generated pn-buildlist-licenses.txt.

We would need a separate script on top of this change and maintain a
database which will map our BPN to the exact source name in SVM
wherever a discrepancy may arrive.

JIRA: SB-22425

Signed-off-by: Tariq Mehmood <tariq.mehmood.ext@siemens.com>