flex_utils/dump_licenses: add check for file existence
We get FileNotFoundError with 'r' mode if we make the first
call to the function with --append. We have such a use case in
jenkins_mel script when we call build_graph() first for the external
toolchain and then again for the internal toolchain. Therefore, before
reading the file, check if it exists otherwise simply go to the next
block where a write operation is performed.

JIRA: SB-22425

Signed-off-by: Tariq Mehmood <tariq.mehmood.ext@siemens.com>