setup-flex-builddir: exclude flex-vendor-integration for qemu
The flex-vendor-integration layer gets included in the build if it is
available because meta-sokol-flex recommends it. No recipe is reachable
in this layer for qemu machines and bitbake complains about it by giving
the following warning:

"No bb files in default matched BBFILE_PATTERN_flex-vendor-integration"

Since this layer is not needed for qemu, we explicitly remove it from
setup-flex-builddir script so that setup-env does not include it while
setting up a build dir.

JIRA: SB-22633

Signed-off-by: Tariq Mehmood <tariq.mehmood.ext@siemens.com>