setup-checkout: fix bug in showing only the latest version of manifest
The implementation was original written assuming that the release
version will be of the type <product>-<version>-<machine> where the
product name will not have any dash (-) character. Since we are using
either sokol-flex or flex-os, the implementation needs to be updated.

Updated the implementation to first use sed to replace the string
`-<version>-` into `\t<version>\t`. Then sorted on basis of version and
then brought back dashes (-).

NOTE: The updated implementation assumes that the version will conform
to semantic versioning format i.e. `[0-9]+.[0-9]+.[0-9]+`. If this
changes in future, the implementation needs to be updated accordingly.

JIRA-ID: SB-23356

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>