distro: replace vendor integration with a simpler implementation
We don't work on vendor BSPs anymore, but we do need a mechanism to
override distro configurations from the BSP layer e.g. to provide
certain packages for the BSP image, or modify PACKAGECONFIG for a recipe
and so on. Remove the existing vendor integration layer logic and
include a distro conf snippet from conf/distro/include/flex-distro.conf

Signed-off-by: Ahsan Hussain <ahsan.hussain@siemens.com>