qemu-system-native: fix mouse co-ordinates scaling issue
This is in continuation of commit 59cc33db.

The previous commit solved the issue with SDL,GL but since scaling
issue was only related to opengl, the scaling shouldn't have been
applied otherwise. So, conditionally scale the mouse co-ordinates
only for SDL,GL.

Moreover, moved the logic to common place and now the scaling isn't
actually required, if we just choose the correct surface dimension
based on opengl context.

JIRA-ID: SB-23711

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>