common/classes: ensure add_sdk_extra_vars is actually executed
This is in continuation of c2086302 where this issue originates.

Either the sdk_extra_vars.bbclass should be directly included in
distro conf file or the function add_sdk_extra_vars should be added
to SDK_POSTPROCESS_COMMAND in the bbclass that is directly included
in distro conf such as sdk_codebench_metadata. This fix will ensure
that environment variables are added to the SDK so that SDK is
visible in CodeBench.

Also bring back the option: `CB_MBS_OPTIONS[general.yocto.sdk.value]`

JIRA-ID: SB-23686

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>