archive-release: drop TEMPLATECONF packaging for installer
TEMPLATECONF that is being packaged for installer is very similar
to the one available in this repository. The separate TEMPLATECONF
for installer was introduced, so that the same machine configuration
including layers information and local.conf, can be saved for which
the installer images are created. There is no need to save-off the
configuration now, so the TEMPLATECONF can be dropped from the
installer.

JIRA-ID: SB-23870

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>