Enable Flex OS builds on RedHat-8 and Rocky-8
Replaced check on /etc/redhat-release with `lsb_release` as
`lsb_release` is available on all DISTROs and is convenient for
setup-environment script.

On Redhat/Rocky-8, at least GCC9 is required. The std::filesystems
library in older GCC versions was located at non-standard paths
and needed to be included manually for multiple recipes. The cleaner
solution is to perform the build with GCC9 and hence gcc-toolset-9
is required.

JIRA-ID: SB-24005

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>