llvm: set LLVM_HOST_TRIPLE for cross-compilation
The correct way to configure cmake for cross-compilation includes
setting the LLVM_HOST_TRIPLE as well.

Documentation Ref:
https://github.com/llvm/llvm-project/blob/llvmorg-19.1.6/llvm/docs/HowToCrossCompileLLVM.rst#configuring-cmake

This fixes the bug when LLVM is cross-compiled for AAarch64 but it
was getting built with the triple of native build system instead of
the TARGET_SYS ('aarch64-oe-linux' in my case). The following error
was observed:
```
No available targets are compatible with triple "x86_64-unknown-linux-gnu"
```

JIRA-ID: SB-24074

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>