mesa: fix vkmark unrealistic score for CPU-based rendering
Port back the patches to replace aemu library provided allocator
`android::base::SubAllocator` to `u_mm.h` allocation APIs.

The vkmark reported score for host CPU-based lavapipe gfxstream ICD
before this change was ~8000 and after this change it is ~1500.

The latter score makes sense as the GPU-based gfxstream ICD gives
a score of ~3500.

Ref: https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/32015

JIRA-ID: SB-24465

Signed-off-by: Haseeb Ashraf <haseeb_ashraf@mentor.com>